+++ app/src/main/java/com/pixlee/pixleeandroidsdk/ui/gallery/GalleryClickListener.java
@@ -1,4 +1,4 @@
-package com.pixlee.pixleeandroidsdk;
+package com.pixlee.pixleeandroidsdk.ui.gallery;
 
 import com.pixlee.pixleesdk.PXLPhoto;
