+++ app/src/main/java/com/pixlee/pixleeandroidsdk/ui/gallery/ListViewHolder.java
@@ -1,4 +1,4 @@
-package com.pixlee.pixleeandroidsdk.gallery;
+package com.pixlee.pixleeandroidsdk.ui.gallery;
 
 import android.view.View;
 import android.widget.ImageView;
@@ -9,11 +9,11 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.pixlee.pixleeandroidsdk.R;
 
 
-public class GridViewHolder extends RecyclerView.ViewHolder {
+public class ListViewHolder extends RecyclerView.ViewHolder {
     public TextView title;
     public ImageView netImg;
 
-    public GridViewHolder(View view) {
+    public ListViewHolder(View view) {
         super(view);
         title = (TextView) view.findViewById(R.id.title);
         netImg = (ImageView) view.findViewById(R.id.netimg);
