+++ app/src/main/java/com/pixlee/pixleeandroidsdk/ui/gallery/PXLProductView.java
@@ -1,4 +1,4 @@
-package com.pixlee.pixleeandroidsdk.gallery;
+package com.pixlee.pixleeandroidsdk.ui.gallery;
 
 import android.content.Context;
 import android.content.Intent;
