+++ app/src/main/java/com/pixlee/pixleeandroidsdk/ui/viewer/ImageViewerFragment.java
@@ -1,4 +1,4 @@
-package com.pixlee.pixleeandroidsdk.viewer;
+package com.pixlee.pixleeandroidsdk.ui.viewer;
 
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
@@ -14,7 +14,7 @@ import com.bumptech.glide.load.DataSource;
 import com.bumptech.glide.load.engine.GlideException;
 import com.bumptech.glide.request.RequestListener;
 import com.bumptech.glide.request.target.Target;
-import com.pixlee.pixleeandroidsdk.BaseFragment;
+import com.pixlee.pixleeandroidsdk.ui.BaseFragment;
 import com.pixlee.pixleeandroidsdk.R;
 import com.pixlee.pixleeandroidsdk.config.GlideApp;
