+++ app/src/main/java/com/pixlee/pixleeandroidsdk/ui/viewer/VideoViewerFragment.java
@@ -1,4 +1,4 @@
-package com.pixlee.pixleeandroidsdk.viewer;
+package com.pixlee.pixleeandroidsdk.ui.viewer;
 
 import android.media.MediaPlayer;
 import android.os.Bundle;
@@ -15,7 +15,7 @@ import androidx.lifecycle.Lifecycle;
 import androidx.lifecycle.LifecycleEventObserver;
 import androidx.lifecycle.LifecycleOwner;
 
-import com.pixlee.pixleeandroidsdk.BaseFragment;
+import com.pixlee.pixleeandroidsdk.ui.BaseFragment;
 import com.pixlee.pixleeandroidsdk.R;
 
 import butterknife.BindView;
